Implementación de cabeceras y modificación en makefile
